Move integration test to `tests/`
Fixes #369.
